src/services/schedule_fetcher.py
"""Schedule fetcher for CDL official schedule page"""
import re
from datetime import datetime
from typing import List, Optional
import requests
from bs4 import BeautifulSoup

from ..storage.models import Match
from ..utils.logger import setup_logger
from ..utils.timezone import to_utc

logger = setup_logger(__name__)


class ScheduleFetcher:
    """Fetcher for CDL official schedule page"""
    
    def __init__(self, schedule_url: str = "https://callofdutyleague.com/en-us/schedule"):
        """
        Initialize schedule fetcher
        
        Args:
            schedule_url: URL of the CDL schedule page
        """
        self.schedule_url = schedule_url
    
    def fetch_schedule(self) -> List[Match]:
        """
        Fetch and parse matches from the CDL schedule page
        
        Returns:
            List of Match objects
        """
        try:
            logger.debug(f"Fetching schedule from {self.schedule_url}")
            response = requests.get(self.schedule_url, timeout=30)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            matches = self._parse_html(soup)
            
            logger.info(f"Parsed {len(matches)} matches from schedule page")
            return matches
        
        except requests.RequestException as e:
            logger.error(f"Failed to fetch schedule page: {e}")
            return []
        except Exception as e:
            logger.error(f"Error parsing schedule page: {e}")
            return []
    
    def _parse_html(self, soup: BeautifulSoup) -> List[Match]:
        """
        Parse HTML to extract matches
        
        Note: This is a basic implementation. The actual HTML structure
        may need to be adjusted based on the real page structure.
        
        Args:
            soup: BeautifulSoup object of the schedule page
        
        Returns:
            List of Match objects
        """
        matches = []
        
        # Try to find match elements
        # Common patterns: match cards, schedule items, etc.
        # This is a placeholder - actual implementation depends on page structure
        
        # Look for common match container patterns
        match_containers = (
            soup.find_all('div', class_=re.compile(r'match|schedule|game', re.I)) +
            soup.find_all('article', class_=re.compile(r'match|schedule|game', re.I)) +
            soup.find_all('li', class_=re.compile(r'match|schedule|game', re.I))
        )
        
        if not match_containers:
            # Fallback: look for any elements with team names
            logger.warning("Could not find match containers, attempting fallback parsing")
            return self._fallback_parse(soup)
        
        for container in match_containers:
            match = self._parse_match_container(container)
            if match:
                matches.append(match)
        
        return matches
    
    def _parse_match_container(self, container) -> Optional[Match]:
        """Parse a single match container element"""
        try:
            # Extract team names
            team_elements = container.find_all(text=re.compile(r'vs|v\.|@', re.I))
            if not team_elements:
                # Try finding team name elements
                team_elements = container.find_all(class_=re.compile(r'team|name', re.I))
            
            # Extract time
            time_elements = container.find_all(class_=re.compile(r'time|date|start', re.I))
            
            # This is a placeholder - actual parsing depends on HTML structure
            # For now, return None to indicate we couldn't parse
            logger.debug("Match container found but parsing not fully implemented")
            return None
        
        except Exception as e:
            logger.debug(f"Error parsing match container: {e}")
            return None
    
    def _fallback_parse(self, soup: BeautifulSoup) -> List[Match]:
        """
        Fallback parsing method when standard patterns don't work
        
        This method attempts to find matches using more generic patterns
        """
        # If the page is JavaScript-rendered, we may need Selenium
        # For now, return empty list and rely on ICS feeds
        logger.warning("Schedule page parsing not fully implemented - relying on ICS feeds")
        return []

